route: Add new netip members to Route{} object
Introduce new members to the Route{} object:

- DstPrefix (type netip.Prefix), which is intended to replace the
  existing Dst (type *net.IPNet) field.

- SrcAddr (type netip.Addr), which is intended to replace the
  existing Src (type net.IP) field.

- GwAddr (type netip.Addr), which is intended to replace the
  existing Gw (type net.IP) field.

These new fields are now included in the Route stringer function, and
for the purposes of this commit will differentiate output depending
on whether these fields carry valid data, or not. The new fields will
take precedent over the old fields.

This commit also updates the Route object comparison function to
include the new fields.

Signed-off-by: Alasdair McWilliam <alasdair.mcwilliam@isovalent.com>