route_linux: Migrate prepareRouteReq() to utilise new fields.
This commit updates the basic serialisation and deserialisation
logic for Route objects to utilise the new DstPrefix, SrcAddr and
GwAddr fields.

To maintain backwards compatibility, this commit introduces translation
logic at the point of data validation, based on the presence of valid
data in the DstPrefix field.

If DstPrefix is valid, we:
  1. Validate we have appropriate SrcAddr and GwAddr data, then
  2. Translate the new fields into the old format and overwrite
     the existing old fields

Otherwise, we:
  1. Validate we have appropriate Src and Gw data, then
  1. Translate the old fields into the new format and overwrite
     the existing new fields

Signed-off-by: Alasdair McWilliam <alasdair.mcwilliam@isovalent.com>